+++ Formatster.Core/TrillionNumberFormatter.cs
@@ -2,7 +2,7 @@ namespace Formatster.Core
 {
     public class TrillionNumberFormatter : INumberFormatter
     {
-        public string Format(long numberToFormat)
+        public string Format(double numberToFormat)
         {
             return (numberToFormat / 1000000000000D).ToString("0.#") + "T";
         }
